feat: MonkeyPatch the reference checks to inject our custom dependencies
Within dbt-core, the _way_ that dbt is invoked will determine if cross-project
references are considered dependencies at a project level. To keep this all simple,
I've implemented a basic monkeypatch that wraps the `is_invalid_protected_ref` and
`is_invalid_private_ref` methods of `Manifest` so that our injected projects are
evaluated as dependencies during access and group checks.